Delete Navbar component
